+++ src/style.css
@@ -32,10 +32,11 @@ ul li {
 }
 
 h2 {
-    text-align: center;
-    margin-top: 5%;
-    font-size: 2.5rem;
+  text-align: center;
+  margin-top: 5%;
+  font-size: 2.5rem;
 }
+
 .hover:hover {
   color: rgb(13, 147, 208);
 }
@@ -103,42 +104,38 @@ footer {
   background-color: rgb(22, 21, 21);
 }
 
-.page-numbers {
-    font-size: 2rem;
-}
-
 .show-controls {
-    display: flex;
-    position: relative;
-    justify-content: center;
-    margin-bottom: 10%;
+  display: flex;
+  position: relative;
+  justify-content: center;
+  margin-bottom: 10%;
 }
 
 .previous {
-    position: absolute;
-    right: 52%;
-
+  position: absolute;
+  right: 52%;
 }
 
 .next {
-    position: absolute;
-    left: 31%;
-
+  position: absolute;
+  left: 31%;
 }
 
 .page-numbers {
-    position: absolute;
-    left: 49%;
-    margin-top: 3%;
+  position: absolute;
+  left: 49%;
+  margin-top: 3%;
+  font-size: 2rem;
 }
+
 .episode {
   display: flex;
   flex-wrap: wrap;
   padding-left: 10%;
+
   /* gap: 1%; */
   text-align: center;
 }
-
 
 .none {
   list-style: none;
